Use float even in int distributions

+++ src/rberga06/phylab/bernoulli.py
@@ -38,7 +38,7 @@ class Bernoulli(DiscreteDistribution):
         return binomial(n, x) * pow(p, x) * pow(1-p, n-x)
 
     @override
-    def p_worse(self, x: int, /) -> float:
+    def p_worse(self, x: float, /) -> float:
         raise NotImplementedError  # TODO: Implement this
 
     @classmethod
